refactor(npm): ignore files for package
