+++ __tests__/models/Route.test.js
@@ -1,10 +1,10 @@
 import { describe, expect, test } from '@jest/globals'
 import { routesTeapot } from './__constants__'
 
-import InvalidArgument from '../../src/errors/InvalidArgument'
-import EmptyMethods from '../../src/errors/EmptyMethods'
+import InvalidArgument from '../../lib/errors/InvalidArgument'
+import EmptyMethods from '../../lib/errors/EmptyMethods'
 
-import Route from '../../src/models/Route'
+import Route from '../../lib/models/Route'
 
 describe('check route configuration', () => {
   test('check invalid configuration', () => {
