+++ lib/__constants__.js
