+++ lib/__typesdef__.js
