+++ lib/errors/EmptyConfigFile.js
