+++ lib/errors/EmptyMethods.js
