+++ lib/errors/EmptyRoutes.js
