+++ lib/errors/InvalidArgument.js
