+++ main.js
@@ -1,14 +1,14 @@
-import EmptyConfigFile from './src/errors/EmptyConfigFile'
-import EmptyMethods from './src/errors/EmptyMethods'
-import EmptyRoutes from './src/errors/EmptyRoutes'
-import FileNotFound from './src/errors/FileNotFound'
-import InvalidArgument from './src/errors/InvalidArgument'
-import InvalidRouteElement from './src/errors/InvalidRouteElement'
-import ModuleNotFound from './src/errors/ModuleNotFound'
+import EmptyConfigFile from './lib/errors/EmptyConfigFile'
+import EmptyMethods from './lib/errors/EmptyMethods'
+import EmptyRoutes from './lib/errors/EmptyRoutes'
+import FileNotFound from './lib/errors/FileNotFound'
+import InvalidArgument from './lib/errors/InvalidArgument'
+import InvalidRouteElement from './lib/errors/InvalidRouteElement'
+import ModuleNotFound from './lib/errors/ModuleNotFound'
 
 export { EmptyConfigFile, EmptyMethods, EmptyRoutes, FileNotFound, InvalidArgument, InvalidRouteElement, ModuleNotFound }
 
-import Parser from './src/Parser'
-import Validation from './src/Validation'
+import Parser from './lib/Parser'
+import Validation from './lib/Validation'
 
 export { Parser, Validation }
