+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fast-custom-router",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Fast custom express router.",
   "type": "module",
   "main": "main.js",
